backend/workers/trends_ingest/requirements.txt
asyncpg>=0.29.0
paho-mqtt==2.1.0
python-dotenv==1.0.1
